Se agrega funcionalidad de consulta cantidad disponible de productos Se modifica en Producto, el tipo de dato cantidad disponible Se agrega un nuevo metodo de verificacion de stock y actualizacion en ProductoService Se rediseña las relaciones de datos, en Venta Se genera en la clase Venta Controller una validacion nueva al crear la venta

+++ src/main/java/com/ventas_bazar/ventasbazar/model/Producto.java
@@ -17,12 +17,12 @@ public class Producto {
     private String nombre;
     private String marca;
     private Double costo;
-    private Double cantidad_disponible;
+    private Integer cantidad_disponible;
 
     public Producto() {
     }
 
-    public Producto(Long codigo_producto, String nombre, String marca, Double costo, Double cantidad_disponible) {
+    public Producto(Long codigo_producto, String nombre, String marca, Double costo, Integer cantidad_disponible) {
         this.codigo_producto = codigo_producto;
         this.nombre = nombre;
         this.marca = marca;
